Add dynamic detection of correct network and disk devices (#272)
* Dynamically determines eth* device with a layer 2 link

Previously, our network config hardcoded eth0 as the network device to
configure. This has worked until today, but is failing with our new Ziply
sites. For the Mellanox cards in the Ziply equipment, the kernel is not
consistently assigning eth* device names to the 6 interfaces on the NIC,
causing network breaks randomly between boots.

This commit introduces a small for loop that looks over all the eth*
interfaces, looking for the one with a layer 2 link, and uses that for the
network configuration.

* Removes hardcoded network interface name from config file

It probably doesn't matter, but the eth* device used may not always be eth0, so
just change the name to 00-mlab.network.

* Dynamically determines disk device ID, rather than using static sda

Rather than just blindly assuming that /dev/sda is the disk for caching and
data, discover an appropriate disk. I found that with the new Ziply sites
/dev/sda is not always the SATA disk, but sometime was showing up with the USB
thumb drive as /dev/sda, which won't work on many levels.

* Fixes for format-cache.sh script

* Removes maxrate specifier from tc command for fq qdisc

Not specifying maxrate means that the maxrate will be unlimited, which should
be equivalent to the uplink speed or the maximum value that can be set for the
value.

* Fixes as bash format error in format-cache.sh

* Sets maxrate parameter of fq qdisc for site with 1g uplinks

* Configures fq qdisc on correct device name

We can't always assume that the device with name eth0 will be the correct one.
Discover the device that has a layer 2 link to the switch.