Removes dep on /dev/sda from format-cache.service
We no longer assume that the main disk will be at device sda, so remove any
explicit dependencies on it in the unit file.

Additionally, with the current configuration I am seeing the format-cache.sh
script fail, because apparently the output of lsblk is not complete, or
possibly even existent when the this unit runs. I try to get past this by
making this unit dependent on sysinit.target, by which time all devices should
have been populated.