Removes the bash -e shopt to avoid script exiting on error
The script runs a loop that checks for a files that may not exist. With -e set,
if it tries to list the files and they don't exist then the script exits. This
commit simply removes the -e shell option. There is likely some other
workaround, but this is pretty simple.