Removes the errexit shell option from generate_network_config.sh (#276)
* Removes the bash -e shopt to avoid script exiting on error

The script runs a loop that checks for a files that may not exist. With -e set,
if it tries to list the files and they don't exist then the script exits. This
commit simply removes the -e shell option. There is likely some other
workaround, but this is pretty simple.

* Changes fq qdisc config script to look for carrier file

The file /sys/class/net/ethN/carrier is what the generate-network-config script
uses, and is probably more appropriate to the intention, and it was just an
oversight that this file didn't previously get changed too.