Store IATA code in /var/local/metadata/iata-code (#277)
* Stores the IATA code in /var/local/metadata on VMs

This can be used by M-Lab-managed VMs to use the Autojoin API. The registration
process requires this value and we need access to it in the register-node
container.

* Corrects bad variable name in write-metadata.sh

* Leverages new loadbalanced VM attribute

Previously, all MIGs were loadbalanced. We are now experimenting with having
VMs use the Autojoin API and removing the loadbalancer. VMs which are
loadbalanced will have the VM metadata field "loadbalanced" set to true,
otherwise false. This also simplifies the write-metadata.sh script.