Updates Cloud Build to use Artifact Registry container images (#281)
* Configures Cloud Build to use Artifact Registry images

Container Registry is being retired next month.

* Changes filesystem path of epoxy_client in build scripts

Seemingly Go's default location for the install command changed from / to /root
with the version of Go in the newer Ubuntu this image is using.

* Use linux grub command instead of linuxefi

linuxefi is not present in /usr/lib/grub/x86_64-efi/ and I found some
information saying to just use linux instead.