Various fixes to the Google OIM VM image builds
Most prominently, I made a mistake with the api_key Packer variable, thinking
it would end up in the environment of the launch-autonode.service script, which
was just wrong. This change set writes the api_key out to a file in user mlab's
home directory, and then reads it in from there when evaluating the Docker
Compose env file.

Additionally, there were some wrong metadata paths and missing curl flags.