Creates dir /opt/mlab/conf in Google OIM VM images
Previously Packer was upload the Google OIM configs to /tmp, but I found that
in the resulting image the files were not in /tmp, even though the build said
it was uploading them there. Perhaps Ubuntu is cleaning out /tmp on boot. To
avoid that possibility the Packer conf now creates /opt/mlab/conf and uploads
all the contents of local ../configs/virtual_google_oim/ to that new remote
directory.

Also, when evaluating the Docker compose environment template, create the file
.env instead of env, which means that we no longer need to specify the flag
--env-file, since .env is the default environment file name for Docker Compose.