Creates "Google OIM" VM images (#284)
* Builds custom VM images for Google-managed Autojoin VMs

We are experimenting with having Google run and manage GCE VMs in a
Google-internal GCP project. These changes add a new Packer build to generate
these custom images. The images should automatically launch the Autonode
services with all the correct settings.

* Adds missing closing quote to string in Google OIM Packer file

* Adds missing comma to array in the Google OIM Packer file

* Renames launch-byos to launch-autonode

Also, removes a cp command for a file that is no longer there.

* Various fixes to the Google OIM VM image builds

Most prominently, I made a mistake with the api_key Packer variable, thinking
it would end up in the environment of the launch-autonode.service script, which
was just wrong. This change set writes the api_key out to a file in user mlab's
home directory, and then reads it in from there when evaluating the Docker
Compose env file.

Additionally, there were some wrong metadata paths and missing curl flags.

* Makes launch-autonode.sh executable

* Creates dir /opt/mlab/conf in Google OIM VM images

Previously Packer was upload the Google OIM configs to /tmp, but I found that
in the resulting image the files were not in /tmp, even though the build said
it was uploading them there. Perhaps Ubuntu is cleaning out /tmp on boot. To
avoid that possibility the Packer conf now creates /opt/mlab/conf and uploads
all the contents of local ../configs/virtual_google_oim/ to that new remote
directory.

Also, when evaluating the Docker compose environment template, create the file
.env instead of env, which means that we no longer need to specify the flag
--env-file, since .env is the default environment file name for Docker Compose.

* Use sudo in Packer config for creating dir /opt/mlab/conf

* Moves config files from /tmp to correct locations in Packer script

This avoid permissions issues and additional Packer configuration blocks, and
consolidates setup in the main VM configuration script that Packer runs as
root.